+++ src/ethereum_test_tools/common/base_types.py
@@ -2,7 +2,23 @@
 Basic type primitives used to define other types.
 """
 
-from typing import Any, ClassVar, SupportsBytes, Type, TypeVar
+from itertools import count, cycle
+from typing import (
+    Any,
+    ClassVar,
+    Dict,
+    Iterable,
+    Iterator,
+    List,
+    Optional,
+    SupportsBytes,
+    Type,
+    TypeVar,
+    Union,
+    get_args,
+    get_origin,
+    get_type_hints,
+)
 
 from pydantic import GetCoreSchemaHandler
 from pydantic_core.core_schema import (
@@ -302,3 +318,169 @@ class HeaderNonce(FixedSizeBytes[8]):  # type: ignore
     """
 
     pass
+
+
+C = TypeVar("C")
+
+
+class ModelIterator(Iterator[C]):
+    """
+    Class that creates model iterators.
+    """
+
+    # Instance fields
+    limit: int
+    iterations: int
+    iterators: Dict[str, Iterator[Any]]
+
+    # Class fields
+    _model: ClassVar[Type]
+    _max_iterations: ClassVar[int] = 1000
+
+    def __init_subclass__(cls, *, model: Type[C], max_iterations: int = 1000) -> None:
+        """
+        Initializes the subclass.
+
+        Model might behave incorrectly if it has a field named `limit`.
+        """
+        cls._model = model
+        cls._max_iterations = max_iterations
+
+    def __init__(self, iterators: Dict[str, Iterator[Any]], limit: int):
+        """
+        Initializes the model iterator.
+        """
+        self.iterators = iterators
+        self.limit = limit
+        self.iterations = 0
+
+    def __next__(self) -> C:
+        """
+        Returns the next model element in the sequence.
+        """
+        if self.limit >= 0 and self.iterations >= self.limit:
+            raise StopIteration
+        if self.iterations > self._max_iterations:
+            raise ValueError(f"exceeded maximum number of iterations ({self._max_iterations})")
+        self.iterations += 1
+        args = {arg: next(iterator) for arg, iterator in self.iterators.items()}
+        return self._model(**args)
+
+
+class ModelGenerator(Iterable[C]):
+    """
+    Class that creates model generators.
+    """
+
+    # Class fields
+    _model_iterator: ClassVar[Type[ModelIterator]]
+    _iterable_fields: ClassVar[List[str]]
+    _index_field: ClassVar[Optional[str]] = None
+    _iter_suffix: ClassVar[str] = "_iter"
+
+    # Instance fields
+    iterators: Dict[str, Iterator[Any]]
+    limits: Iterator[int]
+    current_limit: int
+    current_iterations: int = 0
+
+    @staticmethod
+    def _field_type_is_iterable(field: str, field_type: Any) -> bool:
+        origin = get_origin(field_type)
+        types: List[Any] = (
+            [get_origin(arg) for arg in get_args(field_type)] if origin == Union else [origin]
+        )
+        return any(
+            t is not None and t != bytes and t != str and t != ClassVar and issubclass(t, Iterable)
+            for t in types
+        )
+
+    def __init_subclass__(
+        cls,
+        *,
+        model: Type[C],
+        index_field: Optional[str] = None,
+        max_iterations: int = 1000,
+    ) -> None:
+        """
+        Initializes the subclass.
+
+        Model might behave incorrectly if it has a field named `limit`.
+        """
+        cls._iterable_fields = [
+            field
+            for field, field_type in get_type_hints(model).items()
+            if cls._field_type_is_iterable(field, field_type)
+        ]
+        cls._index_field = index_field
+        cls._model_iterator = type(
+            f"{model.__name__}Iterator",
+            (ModelIterator,),
+            {},
+            model=model,
+            max_iterations=max_iterations,
+        )
+
+    def __init__(self, *, limit: int | Iterable[int] | Iterator[int] = -1, **kwargs: Any):
+        """
+        Initializes the model generator.
+        """
+        self.iterators: Dict[str, Iterator[Any]] = {}
+        for field in self._iterable_fields:
+            field_with_suffix = f"{field}{self._iter_suffix}"
+            if field in kwargs or field_with_suffix in kwargs:
+                if field in kwargs and field_with_suffix in kwargs:
+                    raise ValueError(f"cannot set both {field} and {field_with_suffix}")
+                if field in kwargs:
+                    self.iterators[field] = cycle([kwargs.pop(field)])
+                else:
+                    value = kwargs.pop(field_with_suffix)
+                    assert value is not None, f"value for {field_with_suffix} cannot be None"
+                    assert not isinstance(
+                        value, ModelGenerator
+                    ), f"""
+                        Iter value for {field_with_suffix} in {self.__class__.__name__} cannot
+                        be a ModelGenerator ({value.__class__.__name__}).
+                        Use the normal field name without the suffix ({field}) instead.
+                        """
+                    if isinstance(value, Iterator):
+                        self.iterators[field] = value
+                    elif isinstance(value, Iterable):
+                        self.iterators[field] = iter(value)
+                    else:
+                        raise ValueError(
+                            f"value for {field_with_suffix} must be an iterator or iterable"
+                        )
+
+        if self._index_field:
+            index_value = kwargs.pop(self._index_field, 0)
+            if isinstance(index_value, int):
+                index_value = count(index_value)
+            elif isinstance(index_value, Iterable):
+                index_value = iter(index_value)
+            self.iterators[self._index_field] = index_value
+
+        for arg, value in kwargs.items():
+            if value is not None and isinstance(value, Iterator):
+                self.iterators[arg] = value
+            elif (
+                value is not None
+                and isinstance(value, Iterable)
+                and not isinstance(value, str)
+                and not isinstance(value, bytes)
+            ):
+                self.iterators[arg] = iter(value)
+            else:
+                self.iterators[arg] = cycle([value])
+
+        self.limits = iter(cycle([limit])) if isinstance(limit, int) else iter(limit)
+
+    def __iter__(self) -> Iterator[C]:
+        """
+        Returns an iterator over the elements.
+        """
+        try:
+            current_limit = next(self.limits)
+        except StopIteration:
+            current_limit = -1
+        return self._model_iterator(self.iterators, current_limit)