+++ src/ethereum_test_tools/common/types.py
@@ -49,6 +49,7 @@ from .base_types import (
     Hash,
     HashInt,
     HexNumber,
+    ModelGenerator,
     Number,
     NumberBoundTypeVar,
     ZeroPaddedHexNumber,
@@ -677,6 +678,21 @@ class Withdrawal(WithdrawalGeneric[HexNumber]):
     pass
 
 
+class Withdrawals(ModelGenerator, model=Withdrawal, index_field="index"):
+    """
+    Withdrawals generator.
+
+    Used to generate multiple withdrawals in a test case.
+
+    Takes the same arguments as `Withdrawal`, but if an iterable is provided for a field, it will
+    generate a Withdrawal for each element in the iterable.
+
+    `index` field is optional and will be automatically generated if not provided.
+    """
+
+    pass
+
+
 DEFAULT_BASE_FEE = 7
 
 
@@ -1206,6 +1222,31 @@ class Transaction(CamelModel, TransactionGeneric[HexNumber]):
             if tx.blob_versioned_hashes is not None
             for blob_versioned_hash in tx.blob_versioned_hashes
         ]
+
+
+class Transactions(ModelGenerator, model=Transaction, index_field="nonce"):
+    """
+    Transaction generator.
+
+    This class is used to generate transactions in a test case.
+
+    Takes the same arguments as `Transaction`, but if an iterable is provided for a
+    field, it will generate a transaction for each element in the iterable.
+
+    `nonce` field is optional and if not provided, it will be generated automatically
+    for each transaction starting from zero.
+
+    If none of the values are iterables, an exception will be raised because none of the provided
+    values will be bounded, unless the `limit` argument is provided, which will limit the number
+    of transactions generated.
+
+    Fields that are already supposed to be iterables, such as `access_list` and
+    `blob_versioned_hashes`, can be parametrized as `access_list_iter` and
+    `blob_versioned_hashes_iter` respectively and an iterable of iterables can be provided to be
+    used in each generated transaction.
+    """
+
+    pass
 
 
 # TODO: Move to other file